Modificando Lote y Productos

+++ SistemaVenta.Domain/Entities/Inventory/InventoryMovement.cs
@@ -6,9 +6,9 @@ namespace SistemaVenta.Domain.Entities.Inventory
     {
         public int MovementTypeId { get; set; }
         public int ProductId { get; set; }
-        public int AffectedQuantity { get; set; }
-        public int? OldValue { get; set; }
-        public int? NewValue { get; set; }
+        public decimal AffectedQuantity { get; set; }
+        public decimal? OldValue { get; set; }
+        public decimal? NewValue { get; set; }
 
         public virtual Product? Product { get; set; }
     }
